test: move dsp tests to src
